Descriptor set revamp
